moving get_enterprise_id_for_user to enterprise.api from enterprise.utils
